:bug: fix static scheduling issues with long prompts (#206)
This fixes an issue where batches of requests with long prompts would
not properly be scheduled. Even with a long queue of requests,
smaller-than-full batches would be scheduled because requests would be
rejected from the schedule once the total number of prompt tokens was >=
`--max-num-batched-tokens`.

The `--max-num-batched-tokens` config is designed for chunking up
prompts for chunked prefill, and isn't relevant for static batching.
This PR sets `--max-num-batched-tokens` to the maximum number of prompt
tokens that could be in a full batch, so that the chunked prefill logic
doesn't prevent us from creating large batches.

I missed this before because it requires a lower level test that invokes
the scheduler directly in order to ensure that the full batches are
actually scheduled.

---------

Signed-off-by: Joe Runde <joe@joerun.de>