:arrow_up: Update locked packages (#213)
This PR is the result of running both
```
uv lock
```
to upgrade all the dependencies in the lockfile, and additionally
```
uv lock --upgrade-package aiohttp
```
because there was at least one platform in the lockfile that had a
yanked version locked.

This updates the locked version of vllm to match our pyproject.toml,
from 0.8.5 to 0.9.0.1. This does mean that all of our unit tests will
start using 0.9.0.1 by default so technically we won't be testing
compatibility with 0.9.0. I'm not too concerned at the moment because we
aren't guaranteeing backwards compatibility yet though, and I'd rather
have our docker builds and other locked installs with uv provide the
latest released version of vllm.

Signed-off-by: Joe Runde <joe@joerun.de>