Duplicate the SamplingMetadata class
We were previously reusing the GPU SamplingMetadata
class but there have been incompatible changes upstream
(PR https://github.com/vllm-project/vllm/pull/16728)

Since it's not clear for now whether we want, should
or can reuse the LogitsProcessor implementation as is,
I'm making a copy of the old version of the class for
the spyre backend.

This won't affect any features for now since the vllm
change was an internal refactoring without UX impact.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>