✨ Support prompt logprobs with static batching (#274)
# Description

This PR enables prompt logprobs with static batching, at batch size 1
only. This enables some experimentation and model evaluation tasks on
spyre hardware.

For static batching, the requires us to warm the model up with
`only_last_token=False`, which passes back the hidden state tensors for
the entire (padded) prompt. This is a big performance penalty, and is
also only supported on spyre cards with batch size 1 currently.

So, this PR introduces an environment flag
`VLLM_SPYRE_ENABLE_PROMPT_LOGPROBS` which must be set to 1 to enable
prompt logprobs. At bootup, we check to ensure that we're running in
static batching mode with max batch size 1, and fail to boot otherwise.
All requests that ask for `prompt_logprobs` will be rejected unless
prompt logprobs are enabled. This is different than the behavior today,
where requests always return `[None]` for prompt logprobs.

---------

Signed-off-by: Joe Runde <joe@joerun.de>