Update warmup log messages and comments (#284)
# Description

This came out of a follow up to
https://github.com/vllm-project/vllm-spyre/pull/270 to determine why an
extra Prefill was necessary after using `warmup_mode`. I learned that
the extra Prefill is required to deploy the compiled graph to the Spyre
device. This PR does not change any functionality, but updates logging
and documentation around warmup to make this clearer.

---------

Signed-off-by: Travis Johnson <tsjohnso@us.ibm.com>