Use upstream LogitsProcessors
The changes introduced by PR
https://github.com/vllm-project/vllm/pull/16728
to the sampler architecture were incompatible with
our spyre model runner.

Initially, as a stopgap solution. I copied the old sampling
classes into our vllm_spyre tree just so that we can keep
working on the latest changes from main.

Now this commit reverts that and makes the same logits
processor logic work for the spyre input batch and model
runner classes.  The difference with the gpu model runner is
that in spyre we don't condense the batch but have a boolean
mask that is used to calculate "dense" request indices. These
indices must be used for the BatchUpdateBuilder because they
are the right ones to slice the `logits` tensor that is passed
to the Sampler.

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>