Use VLLM_WORKER_MULTIPROC_METHOD=spawn instead of --forked for tests (#268)
This PR lets us remove the requirement of `--forked` from our pytest
tests.

The hang that is observed without `--forked` is due to the known issue
with libgomp and threading (see [this gcc bug
report](https://gcc.gnu.org/bugzilla/show_bug.cgi?id=58378) that is a
"won't fix"). It is a common problem in Python dueo to the usage of
native libraries behind the scenes. If a process is forked after an
OpenMP thread pool has been created, then the child will not have a
threadpool and the code hangs the next time code enters a parallel
context.


Where this comes up in our tests is actually because we use
`transformers` to compare the generation results. vLLM and PyTorch delay
initializing the thread pool until it is needed. When just using vLLM in
V1, this does not happen in the frontend process, so it is ok to use
`fork()`, but using `transformer`'s `model.generate` in the main process
during the tests initializes the thread pool and causes the next attempt
to create a `vllm.LLM` to hang in the forked worker process.

With `spawn`, the new process is created from scratch and creates a new
OpenMP thread pool. But there are trade-offs here too, eg. using `spawn`
in offline mode requires particular handling in a script. vLLM docs have
a good summary of the trade-offs of this setting in
[REF](https://docs.vllm.ai/en/v0.9.1/design/multiprocessing.html).
Because of that, I just set in in the test environment. Using the `vllm`
cli to run the code actually defaults to `spawn` anyways
([REF](https://github.com/vllm-project/vllm/blob/536fd330036b0406786c847f68e4f67cba06f421/vllm/entrypoints/utils.py#L134-L152)).

FIX https://github.com/vllm-project/vllm-spyre/issues/146

---------

Signed-off-by: Travis Johnson <tsjohnso@us.ibm.com>