:bug: Workaround ray issue in tests (#307)
~~Testing whether or not torch 2.7 will cause the EADDRINUSE failures on
vllm 0.9.2~~

Edit: The issue appears to be with the shutdown logic in vllm, with TP
running the model executor is not shut down when an `LLM` is garbage
collected. This is caused by a circular reference in
vllm.executor.ray_utils when ray is not installed. This PR adds a
temporary workaround to force the engine shutdowns after each test.

Also updating to torch 2.7.1 everywhere since our next release will have
2.7.1 in the base image as well.

---------

Signed-off-by: Joe Runde <joe@joerun.de>