✅ add assertions for warmup mode context (#294)
# Description

This PR adds assertions for operations done within the `warmup context`.
Earlier we saw a bug in which the `decode` warmup was moved outside the
warmup context because of one simple indentation issue and it caused a
segmentation fault with no clue as to why it was happening.

These assertions should catch the error when running even on CPU.

---------

Signed-off-by: Prashant Gupta <prashantgupta@us.ibm.com>