[CB][Tests] Check output of scheduling tests on Spyre (#337)
This code adds some logic to check the ouput of the scheduling steps
tests. The output checking is done only on Spyre to save some compute
for the cpu tests.

[Important note]
The output comparison checking of scheduling tests don't pass on CPU
neither do they on Spyre. Exact reasons why are not sure, on cpu it
might be because of high entropy of the randomly generated prompt
tokens.

---------

Signed-off-by: Sophie du Couédic <sop@zurich.ibm.com>
Signed-off-by: Prashant Gupta <prashantgupta@us.ibm.com>
Co-authored-by: Prashant Gupta <prashantgupta@us.ibm.com>