Upstream compatibility tests (#343)
Since we test against "main" and "default" vLLM versions in our tests,
we often add code to make "main" work ahead of our official support. But
it's easy to forget this compatibility code and it can become technical
debt later. In this PR I'm adding a new test file to that verifies
"main" and "default" imports from vLLM so that the "default" tests
starts failing one a compatibility code is not longer needed.

---------

Signed-off-by: Max de Bayser <mbayser@br.ibm.com>