[CB] Optimization: allowing sequences to join a batch anytime (#340)
### [CB] Optimization: allowing sequences to join a batch anytime

This PR allows to lift the restriction that the prompt length of a new
joining sequence cannot exceed the tkv of the current decode batch. If
all other constraints are met, a sequence can now join anytime,
regardless of the current tkv.

Changes:
- scheduler: check updated constraints for both new joining and current
decoding sequences before adding a new sequence
- model runner: add tkv offset to existing decoding sequences in case a
prompt is longer than the current tkv
- scheduler test: adapting test `test_prompt_too_long_for_current_tkv`

To Do:
- scheduler optimization: depending on the workload, we might want to
optimize over the hold back queue deciding for which sequences to use
this feature.

-> **_update: decision is to enable this feature behind a flag to test
it on realistic workloads._**

closes #346

---------

Signed-off-by: Yannick Schnider <yannick.schnider1@ibm.com>
Signed-off-by: Yannick Schnider <Yannick.Schnider1@ibm.com>