[CB] Support batch size 1 for decode, simplify warmup (#312)
### [CB] Support batch size 1 for decode, simplify warmup

As we moved to torch 2.7.1 in #307 , dynamic dimension of size 1 are
supported by pytorch. Hence, batch size 1 for decode produces the same
graph as batch size >= 2.
This PR relaxes the min batch size 2 constraint for decode and adapts
the warmup. Previously warmup consisted of two prefills and one decode
of batch size 2. The new warmup only features one prefill and one decode
of batch size 1.

---------

Signed-off-by: Yannick Schnider <yannick.schnider1@ibm.com>
Signed-off-by: Yannick Schnider <Yannick.Schnider1@ibm.com>