:zap: cache hf results in tests (#373)
# Description

In an effort to reduce test runtime, this uses `functools.lru_cache` to
cache text generation results from hf transformers. This should shave
about 15% off the cpu runtime based on some quick measurements on an M3.

NB: This probably will not work when running with `--forked`, but we
don't fork the tests on github actions runs.

---------

Signed-off-by: Joe Runde <joe@joerun.de>