Remove block_size from arguments on LLM constructor. (#383)
### Remove block_size from arguments on LLM constructor.

people might be confused by the top level argument `block_size`. We set
it somewhat inconsistently throughout our examples/tests. It has no
influence at all as it gets overridden in any case SB/CB by
[this](https://github.com/vllm-project/vllm-spyre/blob/e6c0d335cad32b65c5f93119dd35ccd25f490fb6/vllm_spyre/platform.py#L188)
line.
For CB this value is the top level arg max_model_len, for SB it depends
on the warmup shapes.

This PR removes it to have more clarity that it does not need to be
passed (and has no impact if passed).

---------

Signed-off-by: Yannick Schnider <yannick.schnider1@ibm.com>