[GHA] 🐛 fix: Save HF models cache for all jobs (#400)
### [GHA] 🐛 fix: Save HF models cache for all jobs

Since #393 , we only load a subset of the cached models instead of all.
However, the condition in "Save HF models cache" assumed that all caches
across the matrix are the same (as previously the case) and only saved
the one for the first job. I removed that condition to save in all
cases.

---------

Signed-off-by: Yannick Schnider <yannick.schnider1@ibm.com>