⚡ Cache LLMs during tests (#396)
# Description

This PR introduces a caching mechanism for LLMs in our tests package.
The rub here is that since we can only have one LLM object instantiated
at a time (ie the cache size is 1), we have to re-order the tests so
that tests with the same parameters are grouped together. This way,
tests that request an LLM with the same config can actually hit the
cache. The test ordering is dumped out to a `.test_sort.txt` file which
prints each test case and it's sort key for easy debugging.

The test re-ordering is handled by `llm_cache.SortKey` and the cache
classes in `llm_cache` handle the actual caching of the LLM resources.
Cache stats are printed at the very end of the test run on a
session-scoped fixture teardown.

There are a few different ways that we boot up vllm engines for testing,
the most common is with `vllm.LLM`. However we also sometimes use the
openai api server, or we construct an engine directly to test the
scheduler. This PR adds caches for `LLM`, `EngineCore`, and remote
openai api servers.

It turns out, for the `eager` backend (which is run here on github
actions) the static batching cases incur a ton of overhead at model
warmup (~10s each), while the continuous batching cases don't (~1s
each). For continuous batching on cpu, the model runtime dominates. This
seems to scale with the # of tokens generated, but also with the KV
cache size in blocks. This PR also includes some test reconfiguration to
reduce the kv cache size and # of tokens generated for many of the cb
tests to bring the runtime down further.

All of this only works if the `--forked` flag is not used

---------

Signed-off-by: Joe Runde <joe@joerun.de>
Co-authored-by: Prashant Gupta <prashantgupta@us.ibm.com>