[cb] scheduler heuristic 2: unblock long prompts (#440)
### [cb] scheduler heuristic 2: unblock long prompts

Introducing `VLLM_SPYRE_MAX_WAITING_TIME_PREFILL`, which is an upper
bound on the waiting time [sec] of any request. After a request has
waited for longer the current decode batch is locked and will finish
decoding. The request will be either added to that locked batch or
prefilled into a new exclusive locked batch.

---------

Signed-off-by: Yannick Schnider <yannick.schnider1@ibm.com>