:bug: implement better checking for granite (#500)
# Fix granite-3.3-8b detection

Fixes #498 

This PR updates the logic that checks for granite-3.3-8b-instruct.
Instead of matching on the model name, it checks the hf config type and
checks for specific values related to the 8b granite variant.

This also consolidates all of the granite-specific overrides to one
method inside `platform.py`. This should make it much easier to both:
- Update these overrides for granite 3.3 8b 
- Add overrides for other models

in the future.

The number of blocks override for the spyre kv cache size was
particularly interesting. We were in a halfway state where we took
overriddes from `--num-gpu-blocks-override` and saved them off to
platform.py and then used the override to change the scheduler's
behavior. I've updated things so that the scheduler no longer needs that
extra override so that we can simply use the configured override
everywhere. This is more robust since the config value is serialized and
given to subprocesses, and more maintainable than managing two separate
numbers for this purpose.

---------

Signed-off-by: Joe Runde <joe@joerun.de>
Co-authored-by: Travis Johnson <tsjohnso@us.ibm.com>