:wrench: Update default simple compile backend (#506)
# Description

Updates the default simple compile backend to `inductor` to match vLLM.

This could be considered slightly risky because it updates default
behavior, but we do want to run with `inductor` where possible. This was
previously set to `eager` because we had issues linking g++ correctly
and needed a fast workaround.

Impact here should be minimal. If users run into problems with
environments that don't have python headers and a valid compiler
toolchain, they can run with `VLLM_SPYRE_SIMPLE_COMPILE_BACKEND=eager`

Signed-off-by: Joe Runde <joe@joerun.de>