[CB] consistent max context length (#514)
We had a (known) inconsistency: Allowing the additional token in
`platform.validate_request`, but not in `scheduler.can_schedule()` (#332
would allow it in the scheduler). However, as the upstream vllm version
vllm-spyre is tied to does not currently allow the additional token,
this PR removes the inconsistency and establishes same behavior as
upstream.

Note: as soon as the upstream version tied to this repo allows the
additional token, we can revert this PR and merge #332 (along with some
unit tests).

cc @sducouedic

Signed-off-by: Yannick Schnider <yannick.schnider1@ibm.com>