Fix test model revision usage (#522)
# Description

Use `ModelInfo.revision` everywhere so that in the future if we need to
use a different version of the model in the tests, we can easily switch
to it.

I tested it by introducing a commit that explicitly disabled the
restored cache, downloaded an older revision of the micro model and the
GH tests passed.

## Related Issues

Fix https://github.com/vllm-project/vllm-spyre/issues/503

---------

Signed-off-by: Prashant Gupta <prashantgupta@us.ibm.com>