:recycle: Simplify env var overrides and add tests (#525)
# Description

~~There seems to be a small bug where passing a valid env var override
will not set the correct value in the environment, only the cache. This
could be an issue where the env var needs to be set to be propagated
across a `fork`, but currently we only use it for updating the batching
mode for poolers so this shouldn't be a big issue atm.~~

Edit: forking would use the cache anyway so this shouldn't be a bug. But
the logic for handling the environment during overrides was a bit funky-
this PR simplifies that code and adds explicit tests covering the
expected behavior.

This PR fixes that, adding some tests to cover our `envs` package. Since
there was an expensive e2e test covering the `sendnn_decoder` backwards
compatibility, I moved it to this new test file so that we only have to
test the config override instead of loading and running a whole model.

---------

Signed-off-by: Joe Runde <joe@joerun.de>