Model runner tests for chunked prefill (#553)
# Description

Tests the new chunking and batching logic at the model runner level.
This should be almost ready to go- I've pointed it into @wallashss's
branch so it can merge into his open PR #552

If we decide to pass back the number of computed tokens from the model
runner, these tests can be updated to take advantage of that. Otherwise
they should be good to review

---------

Signed-off-by: Joe Runde <joe@joerun.de>
Signed-off-by: Wallas Santos <wallashss@ibm.com>