[CP] rewrite scheduler constraints for chunked prefill (🐛 fix) (#560)
rewriting the `can_schedule()` constraints for chunked prefill (reusing
`can_schedule()` of the continuous batching scheduler as currently on
main does not check all conditions correctly).

-> this is a conservative formulation, some conditions (cond3 and cond5)
will be written more optimal soon...

notes:
- for chunked prefill `VLLM_SPYRE_ENABLE_PREFILL_OPTIMIZATION` will
always be **_ON_**.
- getting rid of `VLLM_SPYRE_N_TOKENS_PREFILL_PRIO` for chunked prefill:
does not make sense as we choose the chunk size to be as small as not to
peak ITL too much.

---------

Signed-off-by: Yannick Schnider <yannick.schnider1@ibm.com>