fix in chunked prefill update_states for interleaving (#568)
# Description

There is a bug in `ChunkedPrefillModelRunner`'s `update_state()` wherein
it would not process finished requests in `scheduler_output` when
scheduling a prefill.

The fix in this PR is to still run
`super().update_states(scheduler_output)` in this case but to hide the
prefill request from the `scheduler_output`.

Signed-off-by: Travis Johnson <tsjohnso@us.ibm.com>