[CP] optimal chunked prefill scheduler constraints (#564)
replaces the conservative upper bounds for `cond3` and `cond5` with the
per sequence optimal values for chunked prefill.

changes:
- implements optimal conditions for max model len and batch x tkv
constraints (`cond3` and `cond5`). the latter needed to re-implement the
`check_batch_tkv_limit()` function for chunked prefill.
- scheduler step level test for above conditions (+ max number of blocks
constraint)
- expand engine caching to capture chunk size and add hf cache for new
tests

Note that the implemented test are working for both scheduler policies
with and without interleaving prefills and decodes.

---------

Signed-off-by: Yannick Schnider <yannick.schnider1@ibm.com>
Signed-off-by: Yannick Schnider <Yannick.Schnider1@ibm.com>