Add prefix caching
This implementation uses the KV Cache Manager to find
matching prefixes that satisfy our chunking constraints.
Since in the initial version we're going for scheduler
agnostic caching, the execute_model() function just
executes dummy steps if the chunk was loaded from cache.
This is required because the scheduler controls the
num_computed_tokens.

Signed-off-by: Max de Bayser <maxdebayser@gmail.com>