[PC] Refactor CB model runner to use vLLMs block pool (#585)
# Description

This PR paves the way to better integration with vLLMs KV Cache manager
and also for the reuse of the prefix caching algorithms and data
structures that are already integrated in the BlockPool and the KV cache
managers.

In this PR, the padding blocks are managed outside of the KV cache
manager which will make it easier to take the block mapping directly
from the scheduler in the future.

For now this PR uses the FullAttentionManager which is a
SingleTypeKVCacheManager for simplicity. The KV cache coordinators and
the KVCacheManager are layers of indirection around this which deal with
multiple types of attention and multiple KV Cache Groups. For now, since
we don't support any of this yet I prefer to keep this PR simple.

In general this integration is pretty straigthforward. The only
impedance mismatch that I've found is that the kv cache manager
interface deals with numbers of tokens, while in the spyre model runner,
we sometimes need to work directly with block allocation.

---------

Signed-off-by: Max de Bayser <maxdebayser@gmail.com>
Signed-off-by: Max de Bayser <mbayser@br.ibm.com>