Refactor scheduler tests: split out prompt generation (#606)
# Description

Refactors for tkv step tests. This splits out the prompt creation from
the engine validation, so that we can be very explicit about what
prompts we are constructing for the tests.

IMHO, it makes the tests for these prefix edge cases more
understandable, though it is a bit more verbose for the simple case
where the prompts are just identical.

---------

Signed-off-by: Joe Runde <joe@joerun.de>
Signed-off-by: Prashant Gupta <prashantgupta@us.ibm.com>
Co-authored-by: Prashant Gupta <prashantgupta@us.ibm.com>
Signed-off-by: Joe Runde <joe@joerun.de>