:alembic: Correct the KV cache stats (#609)
# Description

This PR attempts to correct the KV cache stats. This shuttles kv cache
metadata back from the model runner to the scheduler, so that the
scheduler can edit the cache stats before they're sent to the
statloggers. This corrects the stats in the logs and in the /metrics
output.

It would be cool to also patch the logging statlogger so that it says
"Spyer KV cache usage" instead:
```
(APIServer pid=8247) INFO 12-17 13:06:55 [v1/metrics/loggers.py:236] Engine 000: Avg prompt throughput: 0.0 tokens/s, Avg generation throughput: 0.0 tokens/s, Running: 0 reqs, Waiting: 0 reqs, GPU KV cache usage: 0.0%, Prefix cache hit rate: 72.6%
```
but that might require plugging in our own entire logging stat logger
:(.

This _does not_ address the issue that the prefix times will be skewed
when a request is pre-chunked and then queues again waiting on space in
the batch. That optimization is still WIP here:
https://github.com/vllm-project/vllm-spyre/pull/607


This could use some polishing, I'm not sold on reusing the
PrefixCacheStats in the model runner, and there may be a better way to
update these stats than sending more data back from the model runner
that I didn't find. Looking for some feedback before deciding if we
should polish this up and merge or look for a different option.

---------

Signed-off-by: Joe Runde <joe@joerun.de>