Deduplicate prefill (#603)
# Description

When we load blocks from prefix caching, there are two situations where
we need to recompute blocks that are present in cache:
1) The sequences of blocks loaded does not entirely fill a number of
blocks
2) The rightmost blocks are part of the last chunk of the prompt and
must be recomputed regardless of filling this chunk entirely or not

This means that we create duplicates in the block table which
unnecessarily reduce the number of available blocks. This PR
"deduplicates" these blocks by actually running the recomputation on the
padding block 0, by means of pointing the slot mappings of recomputed
blocks to the address range of block 0, but keeping the block table as
if we're loading from cahe.

This PR should also fix a corner case in which the loaded prefix fills
the first K chunks completely, where K is not the last chunk of the
prompt, and where this Kth chunk would be recomputed anyway. (TODO: add
a test for that)

---------

Signed-off-by: Max de Bayser <maxdebayser@gmail.com>
Signed-off-by: Yannick Schnider <yannick.schnider1@ibm.com>
Signed-off-by: Max de Bayser <mbayser@br.ibm.com>
Co-authored-by: Yannick Schnider <yannick.schnider1@ibm.com>